requirements.txt
python-crontab ~= 3.2
sqlalchemy ~= 2.0
psycopg2-binary ~= 2.9
pandas ~= 2.2
requests ~= 2.32
beautifulsoup4 ~= 4.12
google-auth ~= 2.36
google-api-python-client ~= 2.152
pytest-cov ~= 6.0